requirements/common.txt
django==2.2.14
jsonfield==3.1.0
requests==2.22.0
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
django-haystack==2.8.1
django-allauth==0.42.0
django-filter==2.3.0
PyJWT==1.7.1
django-braces==1.14.0
django-ckeditor==5.9.0
django-spaghetti-and-meatballs==0.2.2
django-debug-toolbar==2.2
coverage==5.2.1
pytest-django==3.9.0
sqlparse==0.3.1
django-import-export==2.3.0
flake8==3.8.3
XlsxWriter==1.2.9
faker==1.14.0